move shooting into own script &&\n hideCursor if not debug build &&\n cleanup all files
