+++ Assets/Scripts/CameraFollow.cs
@@ -4,9 +4,9 @@ using System.Collections;
 [AddComponentMenu("Playground/Movement/Camera Follow")]
 public class CameraFollow : MonoBehaviour
 {
-	[Header("Object to follow")]
-	// This is the object that the camera will follow
-	public Transform target;
+    [Header("Object to follow")]
+    // This is the object that the camera will follow
+    public Transform target;
 
     //Bound camera to limits
     public bool limitBounds = false;
@@ -15,57 +15,64 @@ public class CameraFollow : MonoBehaviour
     public float bottom = -5f;
     public float top = 5f;
 
-	private Vector3 lerpedPosition;
-
+    private Vector3 _lerpedPosition;
     private Camera _camera;
 
-    private void Awake() {
+    private void Awake()
+    {
         _camera = GetComponent<Camera>();
     }
 
     // FixedUpdate is called every frame, when the physics are calculated
     void FixedUpdate()
-	{
-		if(target != null)
-		{
-			// Find the right position between the camera and the object
-			lerpedPosition = Vector3.Lerp(transform.position, target.position, Time.deltaTime * 10f);
-			lerpedPosition.z = -10f;
-		}
-	}
+    {
+        if (target != null)
+        {
+            // Find the right position between the camera and the object
+            _lerpedPosition = Vector3.Lerp(transform.position, target.position, Time.deltaTime * 10f);
+            _lerpedPosition.z = -10f;
+        }
+    }
 
 
-
-	// LateUpdate is called after all other objects have moved
-	void LateUpdate ()
-	{
-		if(target != null)
-		{
-			// Move the camera in the position found previously
-			transform.position = lerpedPosition;
+    // LateUpdate is called after all other objects have moved
+    void LateUpdate()
+    {
+        if (target != null)
+        {
+            // Move the camera in the position found previously
+            transform.position = _lerpedPosition;
 
             // Bounds the camera to the limits (if enabled)
-            if(limitBounds) {
+            if (limitBounds)
+            {
                 Vector3 bottomLeft = _camera.ScreenToWorldPoint(Vector3.zero);
                 Vector3 topRight = _camera.ScreenToWorldPoint(new Vector3(_camera.pixelWidth, _camera.pixelHeight));
                 Vector2 screenSize = new Vector2(topRight.x - bottomLeft.x, topRight.y - bottomLeft.y);
 
                 Vector3 boundPosition = transform.position;
-                if (boundPosition.x > right - (screenSize.x / 2f)) {
+                if (boundPosition.x > right - (screenSize.x / 2f))
+                {
                     boundPosition.x = right - (screenSize.x / 2f);
                 }
-                if (boundPosition.x < left + (screenSize.x / 2f)) {
+
+                if (boundPosition.x < left + (screenSize.x / 2f))
+                {
                     boundPosition.x = left + (screenSize.x / 2f);
                 }
 
-                if (boundPosition.y > top - (screenSize.y / 2f)) {
+                if (boundPosition.y > top - (screenSize.y / 2f))
+                {
                     boundPosition.y = top - (screenSize.y / 2f);
                 }
-                if (boundPosition.y < bottom + (screenSize.y / 2f)) {
+
+                if (boundPosition.y < bottom + (screenSize.y / 2f))
+                {
                     boundPosition.y = bottom + (screenSize.y / 2f);
                 }
+
                 transform.position = boundPosition;
             }
-		}
-	}
+        }
+    }
 }
