+++ Assets/Scripts/Player.cs
@@ -8,14 +8,11 @@ using static System.Console;
 [RequireComponent(typeof(BoxCollider2D))]
 public class Player : MonoBehaviour
 {
-    public float shootInterval;
-    public GameObject[] projectile;
     public float playerSpeed = 2.0f;  
     
     private BoxCollider2D boxCollider;
     private RaycastHit2D raycasthit;
     private Vector3 moveDelta;
-    private float _shootTimer;
 
 
     private void Start()
@@ -51,34 +48,5 @@ public class Player : MonoBehaviour
         {
             transform.Translate(moveDelta.x * Time.deltaTime *playerSpeed, 0, 0);
         }
-
-        _shootTimer -= Time.deltaTime;
-    }
-
-
-    public void Shoot(Vector3 mousePos)
-    {
-        if (CanShoot())
-        {
-            Vector3 position = transform.position;
-            float angle = AngleBetweenTwoPoints(position, mousePos);
-            Instantiate(projectile[0], position, Quaternion.Euler(new Vector3(0f, 0f, angle)));
-        }
-    }
-
-    private bool CanShoot()
-    {
-        if (_shootTimer < 0)
-        {
-            _shootTimer = shootInterval;
-            return true;
-        }
-
-        return false;
-    }
-
-    float AngleBetweenTwoPoints(Vector3 a, Vector3 b)
-    {
-        return Mathf.Atan2(a.y - b.y, a.x - b.x) * Mathf.Rad2Deg;
     }
 }
