+++ Assets/Scripts/Shooter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90b080334430cb340a4644f2ac0cd57b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
